requirements.txt
fastapi
pydantic-settings
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
python-multipart
python-jose[cryptography]
bcrypt
jinja2
aiofiles
httpx
websockets
PyPDF2
python-docx
pandas
openpyxl
python-dotenv
Pillow
chardet